Add VolumeIDs for 2011 Ubuntu releases
Note that due to a bug in libosinfo, the detection of 11.10 won't work.
The 11.10 isos don't have a system id nor a publisher id, so they end
up as NULL in the parsed database, but they are a string of N consecutive
space when read from the iso. This prevents the match from working.