Fix setting of property default value
This commit reverts 4ba6a970dd4a6ba5e8efc49bb190c059f0d5b604.
For the default value of a GObject to be set upon object construction,
it needs to be marked as G_PARAM_CONSTRUCT or G_PARAM_CONSTRUCT_ONLY.
This commit adds this flag for the "installer" and "live" properties
("installer" defaults to TRUE and we want this to be set when the
object is created).