Handle NULL GError **
It's valid to pass NULL GError ** around when we are not interested
in error reporting. This means we have to make sure the GError** is
not NULL before dereferencing it. In particular, the right way of
testing if GError **err; is set is to do if (err && *err) {}
This commit adds a helper to do this check and uses it where needed.