Make sure we don't set a GError twice in OSInfoLoader
When parsing OSInfo XML documents, the catchXMLError callback may
be called for non fatal errors. When this happens, we get a pointer
to a parsed document which we can use, but we may have set a GError
during the parsing if a non fatal error was logged. This means we
have to make sure the error is cleared after the parsing, otherwise
we may attempt to set it twice which is forbidden

This can be triggered by running examples/demo.py:

$ python examples/demo.py
/usr/lib/python2.7/dist-packages/gi/types.py:43: Warning: GError set
over the top of a previous GError or uninitialized memory.
This indicates a bug in someone's code. You must ensure an error
is NULL before it's set.
The overwriting error message was: Incorrect root element
  return info.invoke(*args, **kwargs)
Traceback (most recent call last):
  File "examples/demo.py", line 6, in <module>
    loader.process_path("./")
  File "/usr/lib/python2.7/dist-packages/gi/types.py", line 43, in function
    return info.invoke(*args, **kwargs)
gi._glib.GError: at line 18: Namespace prefix xi on include is not defined

Bug report and initial fix from Guido Günther <agx@sigxcpu.org>