Fix "insufficient metadata" test in on_pvd_read
There are 2 issues with this test:
1) it's not using is_str_empty, which means it won't do what's
expected on some ISOs where "empty" descriptors are read as a
string full of blank space
2) it mandates the presence of a system or publisher descriptor in
ISOs, which is not the case on ubuntu 11.10 ISOs. They only have a
volume descriptor. Since on most ISOs, publisher is empty and
system is "LINUX", they don't add much to discriminate between ISOs,
so we don't really need to mandate the presence of one of these.