Mark the volume descriptor structs as packed
This tells gcc it shouldn't add some padding between the struct
members. which is important because we rely on this structure size
being exactly 2048. One of its members is guint8 ignored2[246]; (only
a 2 byte multiple) so all the fields of the struct won't be aligned
on a 8 byte boundary, so gcc could do some unexpected things on
some arch.
The 2 structs changed by this patch only use arrays whose elements have
the same size as chars so gcc will probably don't try to change their
alignment, but one never knows..