Match substring regex in the end
Sometimes volume-id expression of one architecture is a substring of
volume-id expression of another architecture for the same OS. For
example '.*G.*RMC.*' is volume-id expression for i386 media of win7,
while '.*G.*RMC.*X.*' is for x86_64. To avoid incorrect matching, we
need to ensure that matching is done against '.*G.*RMC.*X.*' first and
then '.*G.*RMC.*'.