Add a 'distro' property for OS objects
The OS 'family' property is used to refer to the kernel type of
the OS. The new 'distro' property refers to the grouping of OS
by a particular vendor. For example all Linux distros have a
family of 'Linux', but their own distinct 'distro' value (Fedora,
Ubuntu, etc). Debian is even more fun having alternate releases
which use a Hurd of FreeBSD kernel, these variants will each have
the 'hurd' or 'freebsd' family, with a common 'debian' distro.