Fix Ubuntu 10.04 detection
LTS Ubuntu versions are respinned several times throughout their
lifetime to integrate all the updates on the iso. When this
happens, the volume ID becomes "Ubuntu 10.04.4 LTS i386" for example.
The additional 3rd digit is properly handled for 8.04 but not for
10.04. This commit fixes the 10.04 volume ID regexp so that it handles
this case, and also adds the missing "LTS" to the regexp (rhbz#811871).
This fixes detection of the ubuntu 10.04.4 iso, I couldn't get my hands
on an older 10.04 iso to double check this doesn't cause a regression.