osinfo-detect: ignore DB loading errors
osinfo-detect is currently exiting with a failure when a DB loading
error occurs. However, such errors should not be fatal as they can
happen when the user put a malformed XML file in
~/.local/share/libosinfo. The worse that can happen when this function
fails is that the OsinfoDB will not be populated, but this shouldn't
cause memory corruption or crashes, so we can log the error and go on.