Add 'installer-reboots' parameter to OsinfoMedia
If media is an installer, this specifies the number of reboots the
installer takes before installation is complete. Default value is '1'.

This is mainly needed for applications like GNOME Boxes and virt-install
to be able to tell when OS installation from a given media is complete so
that they can take whatever post installation steps they may need to.