Fix udev rule
The udev rules for detecting OS'es are too complicated and they also don't
work when detecting OS'es on e.g. a loop device. For example, the stuff
about DISK_MEDIA_CHANGE is kinda wrong.

The latter is pretty annoying as the default action in F18 for an .ISO file
is to loopback mount it - as a result the user is presented with "Open with
Files" instead of "Open with Boxes".

Fixes rhbz#863468