InstallConfigParam:policy should not be writable
This actually not only breaks ABI but also the API: we remove one argument
of _new() function. The prop getter is the main part of this API that an
app will be using if its using this API at all so I think its worth it to
correct this now  (while we are causing other breakage here: See next
patch in this series).

I am also not certain that _new() should be part of the public API.