Make 'filename' more consistent w/ other restrictions
'filename' should be toplevel node under 'install-script' rather than
an attribute of its toplevel 'template' child. The reason this was
previously put in the incorrect place was that the first idea was to have
multiple files under one install script but we later decided to go for 1
file per install script.

Also note that I'm naming the element to 'expected-filename'. This is
to make it consistent with associated API and to simplify loading code.