Fix gint64 format string in osinfo_entity_set_param_int64
osinfo_entity_set_param_int64 printed gint64 values using
G_GUINT64_FORMAT which is unwanted when the value is
negative. This bug caused breakage after the changes in
84969ff9 and 4e86e2bf when trying to read the install-reboot value.

Until these commits, we stored the textual representation
of (guint64)-1 (GUINT_MAX) in the entity store, and when trying to convert
it, g_ascii_strtod was returning -GINT64_MAX, and
osinfo_entity_get_param_value_int64_with_default was returning the
default value as this value is negative.

After these commits, g_ascii_strtoll returns GINT64_MAX-1 when trying
to parse the overlong string stored in the database, and the default
value is not returned as the data was found in the database.

Just using the right format (G_GINT64_FORMAT) to convert a gint64 to a
string fixes this bug.