media: Don't mark properties as _CONSTRUCT
Marking them as _CONSTRUCT will cause the property setter to be
called at object construction time to set the property default
value. For the properties of OsinfoMedia, this causes the default
value to be stored in the entity store as the property setters
call osinfo_entity_set_param.

However, as the getters are careful to use
osinfo_entity_get_param_*_with_default, this is not useful,
and wastes memory. This can even be harmful if the default gobject
property value and the default value passed to
osinfo_entity_get_param are different, the gobject default will be
returned, which can be unexpected.