fedora,installer: Set keyboard config for >= F18
Now that we have established what kind of string to expect in
'l10n-keyboard' config, lets correctly set this for F18 and higher.

For F17 and older, we hardcode the layout to 'us'. Those require layout
to be a console layout so we'll need some complicated XSL magic to
translate from X to console layout if we want to properly support those.