Set id in osinfo_install_config_param_new
OsinfoInstallConfigParam is an OsinfoEntity, but its _new method
does not set an id when creating it. This is needed if we want
to be able to use an OsinfoInstallConfigParamList. The "name"
argument that is passed at creation time is expected to be unique,
so we can use that as the entity id even though a full URI would have
been nicer.