loader: Load datamaps
Install scripts can add a 'datamap' attribute when they declare
their config parameters. The value of this attribute is the ID
of a datamap, which is an XML file containing key/value pairs:

<?xml version="1.0" encoding="UTF-8"?>
<libosinfo version="0.0.1">
  <datamap id="http://example.com/osinfo/test">
    <entry inval="generic-val1" outval="foo"/>
    <entry> inval="generic-val2" outval="bar"/>
  </datamap>
</libosinfo>

This commit adds support for loading these datamaps. The next patches
will then make use of these datamaps when the libosinfo user set
a value for the corresponding config parameter.