Formalize architecture names to align with libvirt
The libosinfo database is using a arbitrary unchecked mix of
architecture names. Libvirt has defined a canonical set of
architectures, so copy that across to the RNG schema and then
update the data to match.

The main change is that i386, i486, i586, all merge to just
i686, since in practice these differences haven't mattered
for at least 15 years now.

The 'hppa' arch for Ubuntu images is changed to 'parisc'.