installer,win7: Adapt windows-cmd.xml for Windows 7
On Windows 7, we need to make use of integrated bcdedit.exe to enable
test signing and disable integrity checks before we could attempt to
install (possibly) unsigned drivers. We use the same app to disable test
signing and enable integrity checks after installing drivers.

Since bcdedit.exe does not exist in Window XP, we don't use it on that
OS. Moreover, since the registry key manipulation on Windows 7 does not
help anything, we better not fiddle with those on Windows 7 as doing so
is always discouraged.

Also since the certutil.exe thats used by driver's .cmd file to add
driver publisher to trusted list does not exist on Windows XP, we have to
continue resorting to registry fiddling to enable installation of drivers
from untrusted publishers for Windows XP.

While we are talking about win7, you'll note that the checks are for
vista (6.0 rather than 6.1). This is because the same changes will also
be most probably needed for vista:

http://www.overclock.net/t/187919/how-to-disable-driver-signature-enforcement-in-vista