installer,win7: Require product key
It turns out that windows installer media can have multiple products and
product key is one of the ways for windows to choose which product to
install. In case of Windows 8, unattended installation breaks if a
product is not choosen as part of autounattend.xml file.

Since we already require product key for windows XP and we'll use this
same script for  Windows 8 too, its not too bad to start requiring this
for Windows 7 as well.

Also, the product key will be used to automatically activate the Windows
copy during unattended installation rather than requiring the user to do
it him/herself at most 30 days after installation.

Moreover, the following patch that removes /IMAGE/INDEX, will break
unattended install without a product key in place.

Co-author: Christophe Fergeau <cfergeau@redhat.com>