gnome: Add data for GNOME 3.8
Its mostly just adjusted copy&paste of 3.6 data. The actual live media
is not yet available for 3.8 itself but for 3.7.91/2 they are available
and people have been using them already so better add some info already
so those ISOs are recognized at least.

Before anyone asks, the 3.7.91/2 ISOs use the same volume ID as it will
be on 3.8 ISO.