Move osinfo_install_script_add_config_param to private header
When this method was added, we forgot to add it to the .sym file.
It's meant to work with the osinfo_install_script_*_config_param()
API, but these methods are not really consistent with the rest of
libosinfo API, they belong in an OsinfoInstallConfigParamList class
which was added in 789adb2a.
We could decide to export this method, but since noone should have
been able to use it, it's better to not export it now, and have one
less deprecated method in our ABI.