Add type info to osinfo_loader_entity() 'key' argument
When using 'string(./foo)' as an XPath query, if foo does not exist, an
empty string ("") will be returned. Thus osinfo_loader_string() ignores
empty strings during parsing. This happens during regular parsing as we try
to lookup string(./foo[lang(XX-xx)]) for localized nodes,  and ignore its
result if the XPath query returns ""

However, we want to represent booleans as <foo/>. Boolean values are
currently parsed using osinfo_loader_string(). Parsing <foo/> as a string
would return an empty string. This means such nodes would be ignored as the
current parser is not able to make the difference between an existing node
with empty content and a non-existing node.

By associating type information to the additional keys we want
osinfo_loader_entity() to parse, we'll be able to indicate which values are
booleans, and to parse them accordingly.