build: add libraries via _LIBADD, not _LDFLAGS
Make sure we use the _LIBADD variable for adding our dependencies.

Adding libraries via _LDFLAGS means that we have our libraries appearing
before the user's LDFLAGS environment variable, which is bad, because
the LDFLAGS from pkg-config could contain -L for system paths and
jhbuild sets LDFLAGS to -L${prefix}/lib so that we can find the
libraries in the user's jhbuild install path.  If the system path flags
come before the user's directory then we'll get the wrong version.

Also: automake documentation says it's wrong and would actually error
out on this if it wasn't hidden behind a variable substitution.

https://fedorahosted.org/libosinfo/ticket/12