fedora,rhel: Bigger boot partition
I had a RHEL VM where I got into trouble when yum wasn't able to upgrade
system as my boot partition was too small to fit new kernel image.
Kickstart format allows us to just say that we want the size to be
whatever is the recommended value so lets make use of that.

I tested against RHEL 7.0 and /boot was allocated 512MB.