db: Use volume-size info for detection
If volume-size is known of a media, ensure that it's the same as media
being compared with, for detection. This will enable us to differentiate
ISOs with exactly same volume-id, system-id and application-id.

It's possible that we encounter two different ISOs that have the same
volume-size, not just the other IDs but chances of that happening are
extremely small. Even if that happens, this patch doesn't make things
worse for that scenario.