tests: Pass media to script generator
Now that we have an API to pass media to script generation methods, let's
test that API instead since its not only the recommended API but also this
way we get to test both passing of OS and media information to script
template.

One of the test cases still uses the existing API on purpose to keep the
case of 'no media available' covered as well.