Add osinfo_install_script_generate_for_media*()
Add API that allows apps to pass the media matched against the OS they
are generating the script for. Scripts can use this information to make
decisions about how installation should be setup. For example, we are
going to use this in the Fedora installer script to decide whether to
install from ISO or from network, based on whether or not media is a
network installer.