build: fix VPATH based builds
The VPATH build fails because the enum generator is told to
look for two generated files in the $srcdir instead of builddir.
We should not in fact loom the generated files at all when building
enums, so they can be removed.

The g-ir-scanner is also missing a -I$(build_dir) to let it find
the generated files.

Finally the gtk-doc tools were not being told to look in the
$(build_dir)

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>