loader: rework handling of pci.ids and usb.ids
Currently we have the option to download pci.ids/usb.ids or
symlink to the distro provided version given to configure.

Change configure so that it searches for the files in a
variety of expected locations. This ensures we always use
the external files, in any modern Linux distro, without
the user needing to give an arg. We look for them in

   /usr/share/hwdata/{pci.ids,usb.ids} (RHEL/Fedora)
   /usr/share/misc/{pci.ids,usb.ids}   (Ubuntu/Debian/Gentoo)
   /usr/share/{pci.ids,usb.ids}        (SLES/OpenSuse)

Instead of loading the pci.ids/usb.ids files as part of
the main database file enumeration process, explicitly
load them from their expected location. This avoids the
need to symlink the distro provided files into the database
dir.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>