loader: sanity check entity filenames when loading
When we have loaded the id attribute of the entity, check it
against the entity filename to ensure its XML document was
held in a file with the modern naming.

This merely prints out a warning for now to avoid breaking
existing users, but in the future, files with non-compliant
names will not be loaded at all.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>