loader: fix memory leaks when loading data
The 'lang' attribute needs to be free'd with xmlFree.

The install script config param needs to be unref'd
since when it is added to the install script an extra
ref is taken.

The install script avatar needs to be unref'd since
when it is added to the install script an extra ref is
taken.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>