loader: identify referenced but not defined entities
When loading the database we often have to instantiate
entities to represent relationships (upgrades, derives,
etc) before the entity is actually defined. This makes
the code liable to bugs as we might instantiate entities
which are never defined.

This extends the loader so that it tracks all entities
that are created via references, and once loading is
complete it prints out a warning if any referenced
entities were not defined fully.

This identifies a number of mistakes in our current
database that the following patches will fix.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>