data: add missing Xen 3.1.0 release
The RHEL Xen platforms reference a Xen 3.1.0 release which
was missing entirely.

The 3.1.0 was not officially released as a tar.gz but several
enterprise distros shipped it based off the GIT tag.

Similarly 3.1.1 & 3.1.2 were not officially released. These
are not added since no distro is thought to have used them.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>