list: Fix replacement in osinfo_list_add
OsinfoList maintains a mapping between a const char *entity_id and an
OsinfoEntity object. It does so by using both a GPtrArray and a
GHashTable.

When an entity is added to an OsinfoList, a pointer from the
OsinfoEntity is used as the key (through osinfo_entity_get_id()) and
the OsinfoEntity is g_object_ref'ed and used as a value.
The OsinfoEntity is also appended to the array mentioned previously.
The use of g_hash_table_new_full() ensures that the OsinfoEntity will be
unref'ed when it's removed from the hash table.

However, when trying to add an entity with the same id as one already
present in the OsinfoList, several issues occur:
- g_hash_table_insert() is used, which means the preexisting entity with
  the same id is going to get removed, and thus g_object_unref'ed.
  However, the _old_ key will still be used by the hash table, which
  means that if the OsinfoList was holding the last reference to the
  old OsinfoEntity, the key stored in the hash table is going to
  point to freed memory
- similarly, we make no attempt at removing the preexisting entity from
  the GPtrArray, so this will also contain stale data

This commit makes sure we remove entities with the same id from the
GPtrArray when this situation occurs. It uses g_hash_table_replace()
rather than g_hash_table_insert() which replaces both the key and the
value if the hash table already contains the same key.

This fixes some valgrind issues or crashes that occurred when running
make check in a libosinfo tree with 2 databases, one in the old layout
and one using the new layout.