loader: re-add support for loading files explicitly
Some existing applications pass an XML file to the
osinfo_loader_process_path method, so we must continue
to support this, even though we want apps to switch to
using a directory of XML files, one per entity.

This requires us to re-add the 'skipMissing' flag for
use when loading files.

The other change is that we need to disable the code
that checks relative paths when loading a single file,
instead of files inside a directory.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>