don't include autogenerated enum files in dist
Previous commit 77cf2730eb6e5503a9bd40e6baf7cc0b393a9dc7
reverts 4e488678d4af0e54da7400851d554bc4c4497c76 (fixes
to the VPATH build) because it was said to break the
make distcheck.

After examining the problem, it seems this commit was not
in fact the problem - it merely exposed the existing bug
elsewhere.

The real problem is that we were including the autogenerated
enum files in the dist. So when doing builds from git those
files were in $builddir, but when doing builds from dist
those files were in $srcdir. Except that 'make distclean'
would delete the enum files, which would again cause them
to appear in $builddir.

So the real fix is stop including the enum files in the dist

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>